news-suite/api/requirements.txt
fastapi
uvicorn
apscheduler
readability-lxml
stripe
redis
supabase-py